requirements.txt
fastapi==0.100.0
openai==0.27.8
requests==2.28.1
tiktoken==0.4.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0
